Added option to detect and remove interlace
converter.py is now able to use ffmpeg to detect interlace and apply
detelecine to the process. This is by no means fool-proof so the option
is disabled by default.

It simply looks for ANY frames detected as interlace starting 2min into
the file (to avoid credits) and tests about 100 frames to see if there
are any artifacts,